% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionlessDrawDown.R
\name{AgarwalFunBesselR}
\alias{AgarwalFunBesselR}
\title{The Laplace transforms of Well function}
\usage{
AgarwalFunBesselR(par, Cbar, Skin)
}
\arguments{
\item{par}{variable of Laplace tramsform}

\item{Cbar}{dimensionless welbore storage constant}

\item{Skin}{skin effect factor}
}
\value{
discrete value of Laplace transforms of Well function
}
\description{
Laplace transforms of Well function following the Agarwal  et. al. 1970, based on Bessels functions built in R.
Ideal isotropic medium, slightly compressible fluid and ideal radial flow, wellbore storage and skin effect to Transient flow.
}
\references{
R. G. Agarwal, R. Al-Hussainy and H. J. Ramey Jr.,
“An Investigation of Wellbore Storage and Skin Effect in Unsteady Liquid Flow: I. Analytical Treatment,”
SPE Journal, Vol. 10, No. 3, 1970, pp. 279-290.

Walton, W.C. 2006. Aquifer Test Modeling. Boca Raton, Florida: CRC Press, Taylor & Francis Group.
}
