% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/besselsFuns.R
\name{BesselK0}
\alias{BesselK0}
\title{Approximation of Bessel's functions}
\usage{
BesselK0(x)
}
\arguments{
\item{x}{argumet of Bessel's function}
}
\value{
value
}
\description{
Approximation of Bessel's functions
}
\examples{
x=0.5
BesselK0(x)
}
\seealso{
\code{\link{BesselI1}} and \code{\link{BesselI0}} and \code{\link{BesselK1}}
}
