% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimensionlessDrawDown.R
\name{DimensionlessDrawDown}
\alias{DimensionlessDrawDown}
\title{The dimensionles drawdown of well pumping test}
\usage{
DimensionlessDrawDown(Gwinput)
}
\arguments{
\item{Gwinput}{list of \code{TD} vector of dimensionless time calculated as aquifer Transmisivity[m2/s] * time[s] / well radius [m] / storativity [-], \code{Cbar} dimensionless welbore storage constant,,\code{Skin} skin effect factor}

\item{...}{another agruments}
}
\value{
The dimensionless drawdown calculated for each time of \code{TD}.
}
\description{
The dimensional drawdown of pumping test calcuated using the stehfest algorithm and aquifer parameters.
}
\examples{
# Artifitial exmaple of unconfined aquifer with following parameters of pumping test
# Pumped well discharge rate [m3/s]
Q=0.0045
# Aquifer Transmisivity [m2/s]
Transm=0.012
# Well radius [m]
Rv=1.2
# Storativity [-]
STRR=0.0001
# Time [s]
T = c( 1.88,2.36,2.97,3.74,4.71,5.93,7.46,9.40,11.83,14.89)
# Dimesionles time [-]
TD=T*Transm/Rv/Rv/STRR
# Well drawdown during pumping test [m]
S =c(0.006,0.008,0.010,0.012,0.015,0.019,0.024,0.03,0.038,0.047)
# Dimensionless drawdown [-]
dlS = (S*2*3.14*Transm)/Q
# Aquifer parameters
Cbar= 1562.858133
Skin = 4.976677
#calcutating the dimensionless drawdown using stehfest algorithm
Kytlice_2018_15_05 = PumpingTestDataLoad(Q=Q,Storativity = STRR,
Transmisivity = Transm,Radius = Rv,Time = T,DrawDown = S,WellTestType = "IsotropicMediumBesselR")
Gwinput = SetStehfestData(Kytlice_2018_15_05$DimlesTime,10,Cbar,Skin,WellTestType= "IsotropicMediumBesselR")
simDD <- DimensionlessDrawDown(Gwinput)
# Comparison plot
plot(TD,dlS, pch=19,ylab="Dimensionless Drawdown [-]",xlab = "Dimensionless Time [-]")
lines(TD,simDD,col="red")

}
\references{
R. G. Agarwal, R. Al-Hussainy and H. J. Ramey Jr.,
“An Investigation of Wellbore Storage and Skin Effect in Unsteady Liquid Flow: I. Analytical Treatment,”
SPE Journal, Vol. 10, No. 3, 1970, pp. 279-290.

Stehfest, H. 1970. Algorithm 368 numerical inversion of Laplace transforms. Communication of ACM 13, no. 1: 47–49.

Walton, W.C. 2006. Aquifer Test Modeling. Boca Raton, Florida: CRC Press, Taylor & Francis Group.
}
