% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FittingR.R
\name{GwOptim}
\alias{GwOptim}
\title{The identification of aquifer parameters using inverse problem solution}
\usage{
GwOptim(obs = NA, DimlesTime = NA, OPtFunType = "MSE", ...)
}
\arguments{
\item{obs}{observed dimensionless drwandown}

\item{DimlesTime}{dimensionless time}

\item{OPtFunType}{type of inverse problem minimization based on MSE - \code{MSE}, MAE -  \code{MAE}, MAx error  \code{MAXER}}

\item{...}{additional parameters of \code{RcppDE}}
}
\value{
object
}
\description{
The identification of aquifer parameters using differential evolution
}
\examples{

DimlesTime =c (1.88, 3.74, 11.83,47.1,118.31,470.98,1489.37,3741.12, 14893.66)
obsSnormDra = c( 0.09899987,0.19699975,0.60699923,2.18499723,4.58399418,9.32698816,11.05998596,11.64,12.38)
dd=GwOptim(obsSnorm,dat$T,"MSE",lower=c(0,0),upper=c(200000,200),control = DEoptim.control(itermax=100))

}
