% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PumpingTestDataLoad.R
\name{PumpingTestDataLoad}
\alias{PumpingTestDataLoad}
\title{Pupming Test Data}
\usage{
PumpingTestDataLoad(Q = NA, Storativity = NA, Transmisivity = NA,
  Radius = NA, Time = NA, DrawDown = NA,
  WellTestType = "IsotropicMediumBesselR")
}
\arguments{
\item{Q}{pumping test rate  [m3/s]}

\item{Storativity}{aquifer storativity [-]}

\item{Transmisivity}{aquifer tranmisivity [m2/s]}

\item{Radius}{well radius [m]}

\item{Time}{numeric vector time intervals of puming test [s]}

\item{DrawDown}{numeric vector drawdown recorded for each time of pupmping test [m]}

\item{WellTestType}{the identifier of type of well test , \code{"IsotropicMediumBesselR"} stands for transient radial flow and skin effect to isotropic aquifer with R bessels functions}
}
\value{
list of pumpiming test data
}
\description{
Setting the information about the evaluated pumping test into basic list. Current  Well test seted for stehfest algorithm on Laplace tranform for isotropic medium aquifers with radial flow.
}
\examples{
# Pumped well discharge rate [m3/s]
Q=0.0045
# Aquifer Transmisivity [m2/s]
Transm=0.012
# Well radius [m]
Rv=1.2
# Storativity [-]
STRR=0.0001
# Time [s]
T = c( 1.88,2.36,2.97,3.74,4.71,5.93,7.46,9.40,11.83,14.89)
# Dimesionles time [-]
TD=T*Transm/Rv/Rv/STRR
# Well drawdown during pumping test [m]
S =c(0.006,0.008,0.010,0.012,0.015,0.019,0.024,0.03,0.038,0.047)
Kytlice_2018_15_05 = PumpingTestDataLoad(Q=Q,Storativity = STRR,
Transmisivity = Transm,Radius = Rv,Time = T,DrawDown = S,WellTestType = "IsotropicMediumBesselR")
}
