% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PumpingTestDataLoad.R
\name{SetStehfestData}
\alias{SetStehfestData}
\title{Object for calculating dimensionles drawdaown}
\usage{
SetStehfestData(Time = NA, N = 10, Cbar = NA, Skin = NA,
  WellTestType = "IsotropicMediumBesselR")
}
\arguments{
\item{N}{number}

\item{Cbar}{c bar}

\item{Skin}{skin factor}

\item{WellTestType}{aquifer type}

\item{PumpingTestDataList}{data}
}
\value{
object for calculating Dimensionless drawdown using stehfest algorthm of aquifer \code{IsotropicMediumBesselR} type
}
\description{
Object for calculating dimensionles drawdaown
}
\examples{
# Pumped well discharge rate [m3/s]
Q=0.0045
# Aquifer Transmisivity [m2/s]
Transm=0.012
# Well radius [m]
Rv=1.2
# Storativity [-]
STRR=0.0001
# Time [s]
T = c( 1.88,2.36,2.97,3.74,4.71,5.93,7.46,9.40,11.83,14.89)
# Dimesionles time [-]
TD=T*Transm/Rv/Rv/STRR
# Well drawdown during pumping test [m]
S =c(0.006,0.008,0.010,0.012,0.015,0.019,0.024,0.03,0.038,0.047)
Kytlice_2018_15_05 = PumpingTestDataLoad(Q=Q,Storativity = STRR,
Transmisivity = Transm,Radius = Rv,Time = T,DrawDown = S,WellTestType = "IsotropicMediumBesselR")
Gwinput = SetStehfestData(Kytlice_2018_15_05$DimlesTime,
                         N=10,
                         Cbar= 1562.858133,
                         Skin = 4.976677,
                         WellTestType= "IsotropicMediumBesselR")
}
